moves single station view to its own file
